Move _multi_slice utility into pxla
